Lift common job tracing to new abstract class
